refactor: move user media acquisition logic to UserMediaBuilder
