+++ lib/features/call/utils/utils.dart
@@ -3,5 +3,6 @@ export 'ice_filter.dart';
 export 'sdp_mod_builder.dart';
 export 'sdp_munger.dart';
 export 'sdp_sanitizer.dart';
+export 'user_media_builder.dart';
 export 'video_constraints_builder.dart';
 export 'webrtc_options_builder.dart';
